refactor(content): isolate category ownership
